+++ affinity/cli/commands/reminder_cmds.py
@@ -15,6 +15,7 @@ from ..click_compat import RichCommand, RichGroup, click
 from ..context import CLIContext
 from ..decorators import category, destructive
 from ..errors import CLIError
+from ..mcp_limits import apply_mcp_limits
 from ..options import output_options
 from ..results import CommandContext
 from ..runner import CommandOutput, run_command
@@ -168,6 +169,7 @@ def _validate_single_entity(
 @click.option("--all", "-A", "all_pages", is_flag=True, help="Fetch all pages.")
 @output_options
 @click.pass_obj
+@apply_mcp_limits()
 def reminder_ls(
     ctx: CLIContext,
     *,